v0.10.0 branch merge

+++ k8_cortx_cloud/deploy-cortx-cloud.sh
@@ -161,6 +161,9 @@ buildValues() {
         (.global.storageClass, .consul.server.storageClass) = \"local-path\"
         | .existingSecret = \"${cortx_secret_name}\"" > "${values_file}"
 
+    # Configure all cortx-setup containers for console component logging
+    yq -i '.global.cortx.setupLoggingDetail = "component"' "${values_file}"
+
     # shellcheck disable=SC2016
     yq -i eval-all '
         select(fi==0) ref $to | select(fi==1) ref $from
@@ -641,7 +644,7 @@ function cleanup()
     # to clean up as much as possible going forward.
     # Delete files that contain disk partitions on the worker nodes and the node info
     # and left-over secret data
-    find "$(pwd)/cortx-cloud-helm-pkg" -type f \( -name 'mnt-blk-*' -o -name 'node-list-*' -o -name secret-info.txt \) -delete
+    [[ -d $(pwd)/cortx-cloud-helm-pkg ]] && find "$(pwd)/cortx-cloud-helm-pkg" -type f \( -name 'mnt-blk-*' -o -name 'node-list-*' -o -name secret-info.txt \) -delete
 
     # Delete left-over machine IDs and any other auto-gen data
     rm -rf "${cfgmap_path}"
